Version 1.0
HTML 5 SWIPE GALLERY is a minimalistic html 5 solution to provide tablet
friendly user interface to connect to galleries (subfolders) and images
(image files in sub folders).

INSTALLATION:

Copy all files and subfolders in src folder to your server space and
navigete to this directory (index.php) and the swipe gallery should
function as in demo link give below.

DEMO:

http://www.talaakso.fi/swipegallery/

ADD/SUBSTITUTE OWN GALLERIES AND IMAGES:

Rename the subfolders (galleries) in galleries folder to match your own
galleries.

Delete the old files in gallery folder and transfer your own images to
that specific gallery. Naming convention is imagename.jpg and
imagename_t.jpg for matching thumb image. Images should preprocessed to
sizes: 800px wider dimension for image, 290px for wider dimension for
thumbsimage. For imagenames no whitespaces should be used. Name the
files by convention: Thecla_betulae.jpg. For image names _-characters
will automatically be parsed as whitespaces.

BUGS AND IDEAS FOR FURTHER DEVELOPMENT:

For portrait image there is a annoying flash while scaling down the
image to match ipad2 screen better. This should be easily fixed by
loading the image first to javascript image object and then
onloaded-handler scaling and only there after putting to img-tag object.
In any case it is not visually pleasing to put landscape and portrait
pictures on the same gallery as they just don't occupy the same space.
This version is optimized for landscape images, but could easily be
adjusted to handle portrait images in a better way.

As this is minimalistic demo the layout and code is fixed for
exactly four galleries. This is due to  layout, visuality and usability.
However as galleries and images are searched through for-loops this
could be easily extended to unlimited
amount of galleries and images. Visually this could mean that the
gallery images on the left side could be scrollable by some means.

IN CONCLUSION:

This demo was created to study and create tablet friendly layout (which
should scale automatically to landscape and portrait modes in many iOS
and Android tablets) and user interface which could easily be used with
tablets, laptops and desktops, maybe even smart phones. All this with
minimal coding effort with no need to fix scaling for every device. This
code is meant to provide solid base on which further development can be
based.

Hope you will have fun by using or studying the code.

THIS CODE CAN BE FREELY BE USED FOR ANY LAWFUL PURPOSE INCLUDING
COMMERCIAL USE. THIS DOES NOT INCLUDE THE USE OF MY OWN IMAGES IN IMAGE
GALLERIES WHICH HAVE COPYRIGHTS BY ME. THESE IMAGES MAY ONLY BE USED FOR
NON-COMMERCIAL, EDUCATIONAL AND DEMONSTRATIONAL PURPOSES.

However, if you use my code in your own projects I would be more than
glad to see the results. Also if you have need for a programmer making
these kind of solutions for your project feel free to contact me on my
email.